Fix most broken tests (add in awaits, etc).

+++ tests/conftest.py
@@ -131,6 +131,7 @@ def create_redis(request, event_loop):
     single_connection = request.param
 
     async def f(url: str = request.config.getoption("--redis-url"), **kwargs):
+        flushdb = kwargs.pop("flushdb", True)
         single = kwargs.pop("single_connection_client", False) or single_connection
         url_options = parse_url(url)
         url_options.update(kwargs)
@@ -144,12 +145,13 @@ def create_redis(request, event_loop):
             async def ateardown():
                 if "username" in kwargs:
                     return
-                try:
-                    await client.flushdb()
-                except aioredis.ConnectionError:
-                    # handle cases where a test disconnected a client
-                    # just manually retry the flushdb
-                    await client.flushdb()
+                if flushdb:
+                    try:
+                        await client.flushdb()
+                    except aioredis.ConnectionError:
+                        # handle cases where a test disconnected a client
+                        # just manually retry the flushdb
+                        await client.flushdb()
                 await client.close()
                 await client.connection_pool.disconnect()
